Rename mastr_2.py to mastr.py
